bhyve: Move more amd64-specific code under amd64/
mptable and the e820 are both rather amd64-specific and can be moved
easily.

In the case of e820, move the registration with qemu_fwcfg into e820.c,
as it simplifies bhyverun.c a bit and I can't see any downsides.

No functional change intended.

Reviewed by:	corvink, jhb, emaste
MFC after:	1 week
Sponsored by:	Innovate UK
Differential Revision:	https://reviews.freebsd.org/D40552

(cherry picked from commit 4fe5b70cae6761b9205ff35b72ccfe60d7326301)
