+++ usr.sbin/bhyve/amd64/mptbl.c
