+++ usr.sbin/bhyve/amd64/mptbl.h
